move behavior to behavior from voronoi
